# openapi/matter-manager.yaml
openapi: 3.1.0
...
info:
  title: Matter Manager API
  version: 0.1.0
  description: |
    Thin backend for Matter Manager.

    Responsibilities: authentication, project provisioning, membership, and administration.
    Device data is synchronised browser-to-CouchDB and does not appear here.
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: http://localhost:3000
    description: Local development
  - url: https://api.matter-manager.example
    description: Production
...
security:
  - bearerAuth: []
paths:
...
  /auth/google/callback:
    get:
      tags: [auth]
      summary: Complete Google sign-in
      security: []
      parameters:
        - { name: code, in: query, required: true, schema: { type: string } }
        - { name: state, in: query, required: true, schema: { type: string } }
      responses:
        '302':
          description: Redirect to the web app with a session established
  /auth/token:
    post:
      tags: [auth]
      summary: Issue a CouchDB-validatable access token
      description: |
        Returns a short-lived ES256 (EC P-256) JWT whose `sub` is the CouchDB username. PouchDB sends
        it as a bearer token on replication requests; CouchDB validates it with the public
        key without consulting this API.

        Called on sign-in and again whenever replication receives a 401. Offline clients
        never call it - local writes must never block on token freshness.
      security: []
      responses:
        '200':
          description: A new access token
          content:
            application/json:
              schema:
                type: object
                required: [accessToken, expiresIn]
                properties:
                  accessToken: { type: string }
                  expiresIn: { type: integer, description: Seconds until expiry }
        '401': { $ref: '#/components/responses/Unauthorized' }
...
  /profile:
    get:
      tags: [profile]
      summary: Read the signed-in user's profile
      responses:
        '200':
          description: The profile
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Profile' }
        '401': { $ref: '#/components/responses/Unauthorized' }

    put:
      tags: [profile]
      summary: Update the signed-in user's own settings
      description: |
        Only the fields a user may change about themselves. `sub` and `email` come from the
        identity provider and are not writable here — a profile endpoint that accepted an
        arbitrary `sub` would be an account-takeover primitive.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [locale]
              properties:
                locale:
                  type: string
                  enum: [auto, en, de]
                displayName: { type: string }
      responses:
        '200':
          description: The updated profile
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Profile' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Profile:
      type: object
      required: [sub, email, displayName, locale]
      properties:
        sub: { type: string, description: Internal user id; also the CouchDB username }
        email: { type: string, format: email }
        displayName: { type: string }
        locale:
          type: string
          enum: [auto, en, de]
          description: '`auto` follows the browser; anything else overrides it.'
...
    Problem:
      type: object
      description: RFC 9457 problem details.
      required: [title, status]
      properties:
        type: { type: string, format: uri }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
  responses:
    BadRequest:
      description: The request body is not one this operation accepts
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Unauthorized:
      description: Missing or invalid credentials
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
...